week_04/Assignment_10/requirements.txt
pinecone
openai
python-dotenv